Makefile: explicitly create bin/, if necessary
Most of the go(1) tools will create the missing directory.
Yet, it is better to define this dependency explicitly to make(1),
to avoid unnecessary surprises.

Another useful feature of make(1) is the multiple targets per rule.
This way, we can have the same action and it is the target the variable
of the rule, anymore.