Makefile: split the target from the PHONY rule
A PHONY rule is executed always.  The target of the `readme' rule is to
create the `www/data/readme.json', which depends only on `README.md'.
If we specify these dependencies separately, make(1) won't build the
`readme.json', if it's not necessary.  And yet, we can refer to this
creation by typing `make readme'.